+++ src/Armature/src/SideTuners/Arguments/TuneArgumentRecipient.cs
@@ -3,4 +3,4 @@ using Armature.Sdk;
 
 namespace Armature;
 
-public delegate IBuildChainPattern TuneArgumentRecipient(ITuner tuner, short weight);
+public delegate IBuildChainPattern TuneArgumentRecipient(ITuner tuner, int weight);
